snapalloc: Add support for YV12 in GraphicsConstraintProvider
The graphics lib was not previously queried for YUV formats,
but the alignment was used for YV12 in graphics & video use
cases.
This change adds support in the GraphicsConstraintProvider
so that graphics alignments are factored in for YV12 cases
where graphics texture or render usage bits are set.
This change also updates the default YV12 alignment to 16
to follow gralloc's default alignment.

Change-Id: I02aaaa65d4f87ba51e44806f0403502086394ba0