snapalloc: Fix YV12 plane layout when graphics lib is queried
The current adreno API does not support per-plane
width and height querying.
This lead to incorrect plane layouts for the chroma
planes when graphics constraint provider is enabled.
This change modifies the width and height
to factor in subsampling for YV12 prior to
querying the graphics library.
This change is limited to YV12 as the only
supported YUV formats for the graphics
constraint provider are YV12 and YCbCr_422_I.

Change-Id: I40414b2f12087e1fdb5d28a237d78f97094802f6
(cherry picked from commit 0f6fd91f4328037482b6800a47b695966a7bdc47)