display-core: separate thread for refresh at tui end.
To acquire the VM on the LA side, a commit is needed. If the power is turned off before the commit, the TUI will fail due to a timeout. It's better to run the commit on a separate thread and wait for a set time. If it times out, make all acquire calls for the next commit after powering on.

Change-Id: Ic873924591afbf27005e37c1dbd02118e64b7c3d