display-core: Demura: Enable Demura when strategy changes from default
When there are too many layers that strategy cannot support, it falls to
default strategy that all the layers go to GPU composition. In this case
Demura layer is dropped from layer stack, which causes Demura config is
disabled in driver. When the strategy resumes to support, Demura layer
is staged back into layer stack and submitted to HW. We need to send the
Demura config as well to enable Demura feature back.

Change-Id: I375c254f445dd3501fdc310e62ab70e678388578