sdmclient: Add support to clear buffers mapped to layer
Enable buffer cleanup in SDM client when BufferSlotsToClear
API is invoked.With this support, fbid-to-buffer mappings
associated with layers are also cleared when slot buffers
are invalidated. This ensures that any buffers held or
referenced by the SDM layer corresponding to the cleared
slots are properly released, preventing stale references
and improving memory/resource management.

Change-Id: I1a87c8403282c9755b6cc1fdc795936da57c25ba