display-core: config: update QDCM calib data file for generic panel
This change updates the QDCM color calibration data file for generic
panel on hamoa target. The IGC, PGC and Gamut blobs have been updated to
be compatible with Hamoa DSPP HW blocks thus enabling the features and
the version has been downgraded along with the removal of the common
data.

Change-Id: I3656047235272565306c5cef8e0fa2ef209baa6a